rename: 패키지 이름 소문자로 변경
