+++ MEME-AUTH/src/test/java/org/meme/member/AuthApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class MemberApplicationTests {
+class AuthApplicationTests {
 
     @Test
     void contextLoads() {
