+++ MEME-DOMAIN/src/main/java/org/meme/domain/entity/FavoritePortfolio.java
@@ -1,4 +1,4 @@
-package org.meme.domain.Entity;
+package org.meme.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -7,27 +7,30 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.meme.domain.common.BaseEntity;
 
+
 @Builder
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-public class FavoriteArtist extends BaseEntity {
+public class FavoritePortfolio extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long favoriteArtistId;
+    private Long favoritePortfolioId;
 
     @ManyToOne
     @JoinColumn(name="user_id", nullable = false)
-    private Model model;
+    private User model;
 
-    private Long artistId;
+    @ManyToOne
+    @JoinColumn(name="portfolio_id", nullable = false)
+    private Portfolio portfolio;
 
-    public static FavoriteArtist from(Artist artist, Model model){
-        return FavoriteArtist.builder()
+    public static FavoritePortfolio from(Model model, Portfolio portfolio){
+        return FavoritePortfolio.builder()
                 .model(model)
-                .artistId(artist.getUserId())
+                .portfolio(portfolio)
                 .build();
     }
 }
