+++ MEME-DOMAIN/src/main/java/org/meme/domain/entity/Token.java
@@ -1,4 +1,4 @@
-package org.meme.domain.Entity;
+package org.meme.domain.entity;
 
 import lombok.Builder;
