+++ MEME-DOMAIN/src/main/java/org/meme/domain/repository/ArtistRepository.java
@@ -1,6 +1,6 @@
-package org.meme.domain.Repository;
+package org.meme.domain.repository;
 
-import org.meme.domain.Entity.Artist;
+import org.meme.domain.entity.Artist;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArtistRepository extends JpaRepository<Artist, Long> {
