+++ MEME-DOMAIN/src/main/java/org/meme/domain/repository/FavoriteArtistRepository.java
@@ -1,7 +1,7 @@
-package org.meme.domain.Repository;
+package org.meme.domain.repository;
 
-import org.meme.domain.Entity.FavoriteArtist;
-import org.meme.domain.Entity.Model;
+import org.meme.domain.entity.FavoriteArtist;
+import org.meme.domain.entity.Model;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
