+++ MEME-DOMAIN/src/main/java/org/meme/domain/repository/FavoritePortfolioRepository.java
@@ -1,8 +1,8 @@
-package org.meme.domain.Repository;
+package org.meme.domain.repository;
 
-import org.meme.domain.Entity.FavoritePortfolio;
-import org.meme.domain.Entity.Model;
-import org.meme.domain.Entity.Portfolio;
+import org.meme.domain.entity.FavoritePortfolio;
+import org.meme.domain.entity.Model;
+import org.meme.domain.entity.Portfolio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
