+++ MEME-DOMAIN/src/main/java/org/meme/domain/repository/InquiryRepository.java
@@ -1,6 +1,6 @@
-package org.meme.domain.Repository;
+package org.meme.domain.repository;
 
-import org.meme.domain.Entity.Inquiry;
+import org.meme.domain.entity.Inquiry;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InquiryRepository extends JpaRepository<Inquiry, Long> {
