+++ MEME-DOMAIN/src/main/java/org/meme/domain/repository/ModelRepository.java
@@ -1,6 +1,6 @@
-package org.meme.domain.Repository;
+package org.meme.domain.repository;
 
-import org.meme.domain.Entity.Model;
+import org.meme.domain.entity.Model;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ModelRepository extends JpaRepository<Model, Long> {
