+++ MEME-DOMAIN/src/main/java/org/meme/domain/repository/PortfolioRepository.java
@@ -1,8 +1,8 @@
-package org.meme.domain.Repository;
+package org.meme.domain.repository;
 
-import org.meme.domain.Entity.Artist;
+import org.meme.domain.entity.Artist;
 import org.meme.domain.enums.Category;
-import org.meme.domain.Entity.Portfolio;
+import org.meme.domain.entity.Portfolio;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
